backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.2
yfinance==0.2.38
python-dateutil==2.8.2
openai==1.3.0
google-generativeai==0.3.0
groq==0.4.1
requests==2.31.0
urllib3==2.1.0
alpha-vantage==2.3.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
python-jose==3.3.0
bcrypt==4.1.1
python-multipart==0.0.6
pydantic-settings==2.1.0
